Rename MenuAssets to UiAssets
