+++ src/menu/settings.rs
@@ -1,7 +1,8 @@
 use bevy::prelude::*;
 
-use crate::menu::{button, MenuAssets};
+use crate::menu::button;
 use crate::AppState;
+use crate::UiAssets;
 
 /// Marker for despawning when exiting `AppState::SettingsMenu`
 pub struct StateCleanup;
@@ -12,7 +13,7 @@ pub fn button_exit_settings_menu(In(clicked): In<bool>, mut state: ResMut<State<
     }
 }
 
-pub fn setup(commands: &mut Commands, assets: Res<MenuAssets>) {
+pub fn setup(commands: &mut Commands, assets: Res<UiAssets>) {
     let button_style = Style {
         size: Size::new(Val::Auto, Val::Auto),
         margin: Rect::all(Val::Px(5.0)),
